refactor(rm): move implementation to library
Reflects cargo ops internals
