+++ crates/cargo-rm/src/bin/cargo/commands/rm.rs
@@ -1,53 +1,7 @@
 use cargo::util::command_prelude::*;
-use cargo::CargoResult;
-use cargo_rm::shell_status;
-use cargo_rm::shell_warn;
-use cargo_rm::{manifest_from_pkgid, LocalManifest};
 
-use std::borrow::Cow;
-use std::path::PathBuf;
-
-/// Remove a dependency from a Cargo.toml manifest file.
-#[derive(Debug)]
-pub struct RmOptions {
-    /// Dependencies to be removed
-    crates: Vec<String>,
-    /// Remove as development dependency
-    dev: bool,
-    /// Remove as build dependency
-    build: bool,
-    /// Remove as dependency from the given target platform
-    target: Option<String>,
-    /// Path to the manifest to remove a dependency from
-    manifest_path: Option<PathBuf>,
-    /// Package to remove from
-    pkg_id: Option<String>,
-    /// Don't actually write the manifest
-    dry_run: bool,
-    /// Do not print any output in case of success
-    quiet: bool,
-}
-
-impl RmOptions {
-    /// Get dependency section
-    pub fn get_section(&self) -> Vec<String> {
-        let section_name = if self.dev {
-            "dev-dependencies"
-        } else if self.build {
-            "build-dependencies"
-        } else {
-            "dependencies"
-        };
-
-        if let Some(ref target) = self.target {
-            assert!(!target.is_empty(), "Target specification may not be empty");
-
-            vec!["target".to_owned(), target.clone(), section_name.to_owned()]
-        } else {
-            vec![section_name.to_owned()]
-        }
-    }
-}
+use cargo_rm::ops::cargo_rm::rm;
+use cargo_rm::ops::cargo_rm::RmOptions;
 
 pub fn cli() -> clap::Command<'static> {
     clap::Command::new("rm")
@@ -108,9 +62,6 @@ pub fn cli() -> clap::Command<'static> {
         ])
 }
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-enum UnstableOptions {}
-
 pub fn exec(_config: &mut Config, args: &ArgMatches) -> CliResult {
     let crates = args
         .get_many("dependencies")
@@ -149,48 +100,6 @@ pub fn exec(_config: &mut Config, args: &ArgMatches) -> CliResult {
     };
 
     rm(&options)?;
-
-    Ok(())
-}
-
-fn rm(args: &RmOptions) -> CargoResult<()> {
-    let manifest_path = if let Some(ref pkg_id) = args.pkg_id {
-        let pkg = manifest_from_pkgid(args.manifest_path.as_deref(), pkg_id)?;
-        Cow::Owned(Some(pkg.manifest_path.into_std_path_buf()))
-    } else {
-        Cow::Borrowed(&args.manifest_path)
-    };
-    let mut manifest = LocalManifest::find(manifest_path.as_deref())?;
-    let deps = &args.crates;
-
-    deps.iter()
-        .map(|dep| {
-            if !args.quiet {
-                let section = args.get_section();
-                let section = if section.len() >= 3 {
-                    format!("{} for target `{}`", &section[2], &section[1])
-                } else {
-                    section[0].clone()
-                };
-                shell_status("Removing", &format!("{dep} from {section}",))?;
-            }
-            let result = manifest
-                .remove_from_table(&args.get_section(), dep)
-                .map_err(Into::into);
-
-            // Now that we have removed the crate, if that was the last reference to that crate,
-            // then we need to drop any explicitly activated features on that crate.
-            manifest.gc_dep(dep);
-
-            result
-        })
-        .collect::<CargoResult<Vec<_>>>()?;
-
-    if args.dry_run {
-        shell_warn("aborting rm due to dry run")?;
-    } else {
-        manifest.write()?;
-    }
 
     Ok(())
 }
