+++ crates/cargo-rm/src/cargo/lib.rs
@@ -1,0 +1,20 @@
+//! Show and Edit Cargo's Manifest Files
+
+#![recursion_limit = "256"]
+#![cfg_attr(test, allow(dead_code))]
+#![warn(
+    missing_docs,
+    missing_debug_implementations,
+    missing_copy_implementations,
+    trivial_casts,
+    trivial_numeric_casts,
+    unsafe_code,
+    unstable_features,
+    unused_import_braces,
+    unused_qualifications
+)]
+
+#[macro_use]
+extern crate serde_derive;
+
+pub mod ops;
