+++ crates/cargo-rm/src/cargo/ops/cargo_rm/registry.rs
