+++ crates/cargo-rm/src/cargo/ops/cargo_rm/util.rs
@@ -3,7 +3,7 @@ use std::io::Write;
 pub use termcolor::{Color, ColorChoice};
 use termcolor::{ColorSpec, StandardStream, WriteColor};
 
-use crate::{CargoResult, Context};
+use super::{CargoResult, Context};
 
 /// Whether to color logged output
 pub fn colorize_stderr() -> ColorChoice {
