+++ crates/cargo-rm/src/cargo/ops/mod.rs
@@ -1,0 +1,3 @@
+//! Core of cargo commands
+
+pub mod cargo_rm;
